requirements.txt
fastapi>=0.111
uvicorn>=0.30
pandas>=2.2
pyyaml>=6.0
pytest>=8.2
black>=24.8
ruff>=0.5
mypy>=1.11
